Avoid multiple calls on callback
Please note that this modification is not deeply tested. It was done to solve a SINGLE incorrect behavior, so I don't know if it will still behave normally in other situation.

The problem was that if SVN is not installed/found, `callback` was called on `ChildProcess#error` but also when the process run normally `ChildProcess#close`. Now callback is run only when process leave stdio.

I repeat that this is absolutelly not tested in every case, but it seems coherent.

Also, can you update NPM package to integrate my previous changes & generate documentation?

Thanks